+++ src/accessibility/acObjectList.h
@@ -5,22 +5,13 @@
 #ifndef SHOT_RECORDER_ACCESSIBLE_OBJECT_LIST_H
 #define SHOT_RECORDER_ACCESSIBLE_OBJECT_LIST_H
 
-#include "accessiblefunctions.h"
+#include <QAccessible>
+#include <QString>
 
-// 添加accessible
+class QObject;
+class RecordButton;
+class RecordOptionPanel;
 
-SET_FORM_ACCESSIBLE(QWidget,m_w->objectName())
-SET_BUTTON_ACCESSIBLE(RecordButton, "RecordButton")
-SET_BUTTON_ACCESSIBLE(RecordOptionPanel, "RecordOptionPanel")
-
-QAccessibleInterface *accessibleFactory(const QString &classname, QObject *object)
-{
-    QAccessibleInterface *interface = nullptr;
-    USE_ACCESSIBLE(classname, QWidget);
-    USE_ACCESSIBLE(classname, RecordButton);
-    USE_ACCESSIBLE(classname, RecordOptionPanel);
-
-    return interface;
-}
+QAccessibleInterface *accessibleFactory(const QString &classname, QObject *object);
 
 #endif // SHOT_RECORDER_ACCESSIBLE_OBJECT_LIST_H
